fix: Support OR label filters alongside label MATCH
This fixes an issue where OR label filter expressions would produce
incorrect results when we have both MATCH label and WHERE (... OR ...)
in the same query. For example:

```cypher
CREATE INDEX ON :A;
CREATE INDEX ON :B;
CREATE INDEX ON :C
CREATE (:A:B);
CREATE (:A:C);
CREATE (:A)

MATCH (n:A)
WHERE (n:B OR n:C)
RETURN n;
```

Would return:
```
+--------+
| n      |
+--------+
| (:A:C) |
| (:A:B) |
| (:A)   |
+--------+
```

Also adds some additional tests for edges cases with OR expressions.