Make Docker cleanup more resilient (#3253)
The docker cleanup script does a thorough job in cleaning up after each
CI run, but is susceptible to race conditions with the cleanup jobs that
run on those machines. When the race condition occurs, the script will
exit prematurely and the job will fail.

Improvements:
- changes to the script should make it less susceptible to those race
conditions
- added `always()` to the cleanup steps in the workflow, so it always
runs at the end of the job
- added `continue-on-error` so that, even if this step fails, the job
should be marked as successful