feat: Implement efficient file-transfer mechanism between replicas (#3227)
The PR adds a support for efficient file transfer using our custom RPC
stack. There is a necessary precondition that the file should always be
sent last, after header and request arguments are serialised. The
interface between RPC protocol and message handlers had to be changed so
that the object FileReplicationHandler is also sent. On the builder
side, we are using segments of size 256KiB. When the file is sent, we
first flush the existing segment and then start building a new segment.
On the reader side, we are using a buffer which is at the beginning
64KiB large. If some conditions are met, the buffer can expand but up to
maximum of 4MiB. On the builder side, we use mask 0xFFFF to annotate a
beginning of the stream of bytes which represents a file. A file is
always represented in the same terms:
String marker, filename size, filename, uint marker, file size, file
data. Filename string is validated so that some DoS attacks are
forbidden.
On the reader side, when the file request is received, we cache
header+request as a vector of bytes.