Fix flaky e2e triggers tests (#3288)
The end-to-end trigger tests are rarely failing: [example
1](https://github.com/memgraph/memgraph/actions/runs/17495831469/job/49696364914)
, [example
2](https://github.com/memgraph/memgraph/actions/runs/17731298906/job/50382887488).
I wasn’t able to reproduce this locally, but the logs show that the
database starts executing the trigger query almost at the same time it
receives the query that checks whether the trigger succeeded. Since the
issue is hard to reproduce and the failures always occur in AFTER COMMIT
tests, I don’t think this points to a bug in the trigger logic. It seems
to be just a timing issue.