feat: Add support for routing based on replica lag (#3222)
The PR adds a new coordinator setting `max_replica_read_lag_`. The point
of this setting is to forbid routing read queries to replicas if they
are too far behind the current main.
A new behaviour is also that the follower will forward the
GetRoutingTableRpc to the cluster leader in order to guarantee
correctness of the current routing table at the expense of network
latency call.