Combine testing Python environments (#3271)
This PR should address some of the flakiness that was likely introduced
when `uv` started being used in CI. We use a cache that is shared
between runs to speed up package installation - `uv` makes links to
packages a global cache to the virtual environment and appears to be
occasionally linking to packages of the wrong version.

To fix this:
- Merged all testing environments into a single test env under
`tests/ve3` using a single set of requirements defined in
`tests/requirements.txt`.
- Partially clear the `uv` cache when stopping the build container, such
that only compiled packages should remain. in order for this to be
possible, a small hack has been introduced by setting the `UV_CACHE_DIR`
to a directory within the usual location of `$HOME/.cache/uv`, because
that directory is a bind mount and `uv` cannot clear it.
- Set `export UV_LINK_MODE=copy` when creating the virtual environment
so that `uv` copies packages rather than creating hard links in order to
avoid linking to the wrong package.


This should address some recent failures:

- release stress
https://github.com/memgraph/memgraph/actions/runs/17725545620/job/50365354244
- release stress
https://github.com/memgraph/memgraph/actions/runs/17661842271/job/50200574660
- daily build release durability and stress
https://github.com/memgraph/memgraph/actions/runs/17658478904/job/50186905357