fix: Improve disk usage in HA by eagerly deleting received WAL file on the replica (#3300)
For a brief period of time there will be snapshot_retention_count+1
snapshots in the system because we first create a snapshot and then
delete the oldest one -> important when doing a capacity planning on the
K8s side.

Replica doesn't manage efficiently disk space:
- its WAL files and possibly received snapshots never get deleted (in
normal circumstances, replica up, main up) because the cleaning of WAL
files is done only when the snapshot is created and replicas don't
create snapshots. On replica they get transferred to .old files if
SnapshotRpc or WalFilesRpc is received as part of the force reset
operation.

WAL files get cleared on the current main only when there are exactly
storage_snapshot_retention_count snapshots in the system, including the
current one.

The PR adds the following:
- replica will delete the WAL file it received from the current main.
- found corrupted snapshot won't be deleted anymore
- tested and refactored retention code for WAL files and snapshots
- refactored `GetRecoverySteps` code