Return current clock time when calling `datetime` with just timezone (#3339)
When calling `datetime()` with only a timezone field in the given map:

```cypher
RETURN datetime({timezone:"Europe/Brussels"});
```

this returned a `datetime` with the minimum value for all numeric
fields, such as year, month, day, hour, etc:

```
0000-01-01T00:00:00+00[Europe/Brussels]
```

The expected behaviour is that when *only* a timezone is given,
`datetime` should compute the current clock time in the given timezone.
This PR implements the correct (i.e., same as Neo4j) behaviour, but also
breaks the previous `datetime()` result if anyone was relying on that
specific behaviour.

The same query will now return something like:

```
2025-10-13 11:20:28.174118000[Europe/Brussels]
```