test: Refactor fine_grained_access to target specific revokes
Previously, we had "hidden" dependencies, in that a single method was
used to clean up and revoke anything a test could have granted. This was
a hidden dependency, and it would be easy for someone to add a (broken)
test that would pass because some GRANT from a previous test lingered.
Sorry about the diff noise, but being explicit is much cleaner!