feat: Support loading parquet files from local disk (#3356)
A new query is added `LOAD PARQUET FROM...`. New library `arrow` is
added which is built through `libs` folder functionality. Loading
parquet files uses a prefetcher thread which operates in column batches
and from such column-wise oriented data, assembles rows which the main
execution thread at the end just pulls. That's how we hide the cost of
assembling rows from column-oriented data. To avoid copying allocated
data from the background thread to the main thread, we make use of
thread-safe memory resource. A new data queue is added which is used as
a shared resource between the prefetcher and the main thread.


The class `utils::Date` supports instantiation now using the number of
days and the number of microseconds passed since UNIX epoch.

---------

Co-authored-by: Dr Matt James <mattkjames7@gmail.com>