refactor: Optimize FinalizeWal by avoiding copying (#3381)
When finalising WAL file, single instance can move the WAL file instead
of copying it. Even in HA is this an option unless the path is locked.
When the file is used by a replication process, the copy will be used
instead of move.
This should optimise a lot of MG scenarios, especially when using
durability during import.