Make Ubuntu mirror selection more resilient to network issues (#3422)
Some sort of network issues between the CI machines and the Ubuntu
mirrors appear to be causing the build to fail before it has even
started, likely at the point where `curl` is used to fetch the mirror
list. This change should allow the build to continue using the default
mirror if that happens.