feat: Add support for snapshots on replicas (#3420)
Enables creating snapshots on replicas. The underlying motivation of
this PR is to speed up replica's recovery. When receiving snapshot from
main, we will try to acquire the snapshot lock. If failing to do so, we
try to abort the current snapshot creation and retry acquiring the lock,
prioritising in this way received snapshot from MAIN over locally
created one.