Generate SBOM when packaging (#3454)
When packaging for the daily build, RC build or optionally during manual
packaging, produce an SBOM.

The SBOM is created in 2 or 3 parts:
1. `conan` produces part of the report, containing the libraries it
provides for the build.
2. `libs/` contains a manually produced SBOM for those libraries
provided either by the toolchain, or by `setup.sh`.
3. `syft` produces a third component if a Docker image is created,
including any packages contained inside the image.

Parts 1 and 2 are combined into a single JSON:
`memgraph-build-sbom.json`; if part 3 exists, then that will be combined
with `docker-sbom.json` to produce `memgraph-docker-sbom.json`. All JSON
SBOMs are produced in the CycloneDX 1.6 format.

Finally, `memgraph-build-sbom.json` and, if it exists,
`memgraph-docker-sbom.json` are processed into a human-readable table
within a text file with the same name as the JSON.

The above outputs are uploaded as an artifact called
`memgraph-sbom.zip`.