Force shared lock for GC in analytical mode and add runtime flag for GC aggressiveness (#3448)
The `FREE MEMORY` query was forcing a unique lock on `main_lock_` in
analytical mode, which would block other transactions. This was
frustrating and counterintuitive for users expecting high performance
with `IN_MEMORY_ANALYTICAL`.
Also, a runtime flag `storage-gc-aggressive` has been added to check
whether the GC should be aggressive.