testing: Add Jepsen concurrency parameter (#3489)
We can now set different concurrency parameter when running Jepsen
workloads. Concurrency relates to the number of worker threads used in
Jepsen test. Previously it was hard-coded to the number of nodes.
Fixed concurrency bug in which multiple clients could have entered the
critical section and import the dataset or register instances making the
test invalid immediately at the beginning.