Fix text index recovery from snapshot (#3570)
Currently, text index recovery works only when the text index
directories exist; otherwise, the index is not recovered. This PR fixes
this issue by checking whether the index folder exists and, if it does
not, recovering it from the snapshot.
Additionally, it improves the safety of the Clear() function (which is
called during DROP GRAPH). Previously, this function was not safe
because it could fail during execution and leave the system in an
intermediate state if a failure occurred while clearing.