Several IN_MEMORY_ANALYTICAL improvements and bug fixes (#3461)
There are multiple issues we are addressing in this PR:

1. Create a snapshot when switching from analytical to transactional
mode.
2. Fix a bug where deltas were being created for indexes and constraints
in analytical mode, even though our documentation states that analytical
mode does not generate deltas.
3. Since the text index doesn’t work in analytical mode, prevent users
from creating a text index while in analytical mode.
4. Fix an issue with uniqueness constraints in analytical mode, where it
is currently possible to violate them and end up in a corrupted state
when switching between storage modes.

closes #3436