Fix slow library pulls from git (#3598)
Inside CI, the `libs/setup.sh` script should clone libraries from a
local docker container that hosts mirrors of the git repos of the
external libraries. This is blocked by the firewall, meaning that the
clone would timeout and eventually clone from the original repo ->
resulting in a _very_ slow prep for the build. This PR will add a docker
network that the mgbuild container can access mgdeps-cache from.